pyproject.toml
[tool.poetry]
name = "grow-tower-auto-clicker"
version = "0.1.0"
description = ""
authors = ["Joe Ferner <joe@fernsroth.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.1"
Pillow = "^9.0.0"
opencv-python = "^4.5.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
